Split Timer in to three components so far. Added padding to seconds for display in timer, eg 01 instead of 1
